feat(node::Find): also change `Node::find` to return `node::Find`
Using `Node::find` to fetch only a few nodes should be _much_ faster
as the descendant nodes are checked when required, instead of being
checked upfront in the old implementation which returned a `Selection`.

Only counting the nodes matching a predicate should also be faster than
before due to no dynamic memory allocation in `Node::find` and
`node::Find`.

Fixes #19.